Move shared models from view_model to model
