+++ xyz-iinuwa-credential-manager-portal-gtk/src/model.rs
@@ -1,0 +1,81 @@
+use serde::{Deserialize, Serialize};
+
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Credential {
+    pub(crate) id: String,
+    pub(crate) name: String,
+    pub(crate) username: Option<String>,
+}
+
+#[derive(Debug)]
+pub enum CredentialType {
+    Passkey,
+    // Password,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct Device {
+    pub id: String,
+    pub transport: Transport,
+}
+
+#[derive(Debug)]
+pub enum Operation {
+    Create { cred_type: CredentialType },
+    Get { cred_types: Vec<CredentialType> },
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub enum Transport {
+    Ble,
+    HybridLinked,
+    HybridQr,
+    Internal,
+    Nfc,
+    Usb,
+}
+
+impl TryInto<Transport> for String {
+    type Error = String;
+
+    fn try_into(self) -> Result<Transport, String> {
+        let value: &str = self.as_ref();
+        value.try_into()
+    }
+}
+
+impl TryInto<Transport> for &str {
+    type Error = String;
+
+    fn try_into(self) -> Result<Transport, String> {
+        match self {
+            "BLE" => Ok(Transport::Ble),
+            "HybridLinked" => Ok(Transport::HybridLinked),
+            "HybridQr" => Ok(Transport::HybridQr),
+            "Internal" => Ok(Transport::Internal),
+            "NFC" => Ok(Transport::Nfc),
+            "USB" => Ok(Transport::Usb),
+            _ => Err(format!("Unrecognized transport: {}", self.to_owned())),
+        }
+    }
+}
+
+impl From<Transport> for String {
+    fn from(val: Transport) -> Self {
+        val.as_str().to_string()
+    }
+}
+
+impl Transport {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Transport::Ble => "BLE",
+            Transport::HybridLinked => "HybridLinked",
+            Transport::HybridQr => "HybridQr",
+            Transport::Internal => "Internal",
+            Transport::Nfc => "NFC",
+            Transport::Usb => "USB",
+        }
+    }
+}
